StagefrightRecorder: avoid 30ms offset with QC AAC encoder
There is a loss of a/v sync during camcorder recording with
QC HW AAC encoder when pause/resume operations are done continuously.

This is because the encoder does not consider the input timestamp
to calculate the timestamp of an encoded frame. For each output buffer,
the timestamp is just the number of frames encoded so far times the frame
duration.
However, as the input timestamp is used by video encoder, there is a drift
between the audio and video timestamps after resume which results in a/v sync loss.

Avoid the default 30ms offset in the total paused duration when QC AAC
encoder is used to ensure that the timestamps of audio and video are in sync.

Also, calculate the TotalPausedDuration value considering the
capturerate and framerate.
Change-Id: I8420e44ab96484f0d6301c366a24eefc8efeaf0f

CRs-Fixed: 1070540
Change-Id: Ia201f3df011c61ce8ed47c3ccb7a2e6a457c2cef