Update patch set 1
Patch Set 1:

> Set the flag in vendor head or elsewhere rather than creating new
 > flags.

I still think this is the better and cleaner and better way to go:

We are including head-capabilities here:
https://github.com/AOSPA/android_vendor_pa/blob/nougat-caf/main.mk#L134
This includes the flag which indicates that vendor/head is present. It wouldn't be good to globally enable qti resampler when vendor/head is present since it needs device specifc tests and depends on other proprietary things. I would prefer this solution pretty much.

Patch-set: 1
Label: Code-Review=0