audioflinger: remove redundant adding to tracksToRemove
There could be a corner case when direct track paused just after
started before rendering anything and HAL is still in standby.
So this active track can be added to tracksToRemove by two times,
then it is started again before threadloop_removeTracks() called.
And finally when threadloop_removeTracks() is invoked, it will
trigger two times stopOutput and makes the active count of this
stream to 0.
Remove the redundant adding where it can also be added when track
is paused and all frames are presented out.

Change-Id: I9f7575a677a25a348dce750f8b338f45646d309c
CRs-Fixed: 1091752