soundtrigger: Support multiple clients attaching to a module
Currently service allows only a single client to be attached
to a Module.
This limits only a single client can use sound trigger at a time.
Add changes to attach multiple clients to a given Module through
ModuleClient inteface so that multiple clients can in parallel
use sound trigger on a given Module. ModuleClient class is
introduced as a client interface to a Module.
Service provides a unique instance of ModuleClient to each client
being attached and adds this client to the module clients list.

Bug:32030191
Change-Id: I0214395ed545690bb9f95365c7b8d415fcf60c9a