audio: Output latency update in Audio Track
Audio Track needs to query latency from the output descriptor each
time it uses mAfLatency to calculate timestamp or return track
latency. Until now, the output descriptor latency is being queried
and stored when a track is created. However there are usecases like
when there is a device switch but not a track tear down, the path
latency changes but the track does not requery the output latency.
This fix updates the output latency in Audio Track such that the
latency being used reflects the correct device latency.

Change-Id: Ided33bbb0daa8ae43f0af169cdc423b204f981f0
CRs-Fixed: 2004382