audio: resize resampler's inframe buffer independent of out buffer
As sample rate of input frame may change dynamically, it is possible
that the input frame size requested may grow given the output frame
size is fixed. If we don't enlarge the input buffer in this case,
array out of boundary occurs and corrupts the heap.

Fix is to resize resampler's inframe buffer independent of out buffer.

CRs-Fixed: 2044381
Change-Id: I43f70dbb3eaadabc8ed06817a04165feab320b3c