Update patch set 1
Patch Set 1:

> I think the BOARD_NEEDS_MEMORYHEAPION in Android.mk is obsolete.
 > 
 > Also which devices would need TARGET_CAMERASERVICE_CLOSES_NATIVE_HANDLES?

Mi5 has that set as well, but it uses HAL3 and this isn't a problem for it. Fabian, disable TARGET_CAMERASERVICE_CLOSES_NATIVE_HANDLES on gemini if this is merged if you don't need it.

Patch-set: 1
Reviewer: Gerrit User 1000009 <1000009@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Verified=0