nuplayer: Fix playback speed issue of video only clips
When playback rate changes after some time, Renderer assumes
AnchorTime is also updated to the same time, but for video
only clips AnchorTime is zero since it is updated only in the start.
Due to this there is a delay in rendering the buffers. Hence
clearing Anchortime while playback rate change, This corrects
AnchorTime while rendering video data.

CRs-Fixed: 2160581
change-Id: I0ef492ca2b8c8dfbbc103a16cb48fee82adaa589