audiopolicy: Allow channel mask exact match for record use case
Channel mask choosing policy for record sessions are flexible
even if the profile does not have the exact channel mask for
the requested usecase. Change this policy to do a exact match
for channel masks as requested so that multi channel use case
does not take the wrong profile.

Change-Id: If435523de48653a4e1e3aa551b93b90a5ac6d662
CRs-Fixed: 2332860