audiopolicy: add device switch handling for WFD AFE PROXY playback.
WFD audio routes back to target speaker when starting a ringtone
playback. This is because primary output is reused for ringtone,
so output device is updated based on SONIFICATION strategy for
both ringtone and music playback.The same issue is not seen on
remoted_submix HAL based WFD audio because primary output is not
reused and a new output is created for ringtone playback.
Issue is fixed by updating output flag to AUDIO_OUTPUT_FLAG_FAST
when there is a non-music stream playback on WFD,
so primary output is not reused for ringtone.

CRs-Fixed: 2864976

Change-Id: Ie80387e3a0f417c751fe68bde27ea3849e7d7585