nuplayer: proper handling of audio start latency for A/V sync
During non-offload playback, when pause and seek are issued, audio
sink needs to be restarted. In this scenario A/V sync was handled
by considering sink start latency. But this has exposed a race
condition during start, when renderer is paused as part of pre-roll.
This is to ensure that above handling will be applicable only for
client-pause usecases.

CRs-Fixed: 3159584
Change-Id: Id8d2a5f4a06fbff16736999e11631f53f167ddea