Spatializer: recenter head pose when head tracking is disabled
Make sure a centered head pose as well as a disabled head tracking
mode is sent to the spatializer effect engine when head tracking is
disabled or the head tracking sensor unregistered.

Previous implementation was relying on updates from the
onActualModeChange() callback from the head pose controller but this
callback is not garantied when the pose controller is destroyed.
Also a neutral head pose is explicitely applied to the engine before
disabling headtracking.

CRs-Fixed: 3364595
Bug: 249820860
Test: repro steps in bug.
Change-Id: I3c8fee7c1cd763206826d4d5f9058abc3231b118
(cherry picked from commit bdecc0545677fc068e26a2d307edbcc8b3151ffc)