libaudiohal@aidl: Fix the implementation of MMap 'start'/'stop'
In the implementation of 'start', consider the current state
of the stream, and apply relevant commands to bring it into
the ACTIVE state.

In the implementation of 'stop', also consider the current
state, and bring the stream into the IDLE state. Previously
the code was driving the stream into STANDBY state, which
is not correct considering that there is a dedicated
StreamHalInterface call for that.

Bug: 335103821
CRs-Fixed: 3786699
Test: atest CtsNativeMediaAAudioTestCases
Change-Id: Ib73bf6d2d4da1a1a146aa63706b61c7aebc897b6
(cherry picked from commit 40618538671cb540f485fec790be6a4e1f020d56)