audiopolicy: Fix direct pcm behavior when 2 direct tracks are played
With new direct_track_reprioritization feature, the latest request of a direct track is catered and previous output is closed. However, if the previous track is still active, then AudioTrack related to this output is invalidated and tries to restart.

For direct pcm tracks, there is no fallback for flags, so if 2 direct tracks are played, this causes continous closing the output
of each other leading to no sound. First track takes pcm_offload path, when second request comes in it closes the existing output, now as output is closed, 1st direct track gets invalidated and restarted with same flags, and it closes the output of second track (which was recently prioritized over first) and this recursive behavior goes on leading to no sound.

Unlike compress offload, in direct pcm, there is currently no logic in client side to not to create direct track, if previous request fails.

Instead, don't prioritize new request if there is already a direct pcm track ongoing.

CRs-Fixed: 3865527

Change-Id: I242c58849e359568fcf1ce8b1b459465892f471a